% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc.R, R/metaboanalyst_general.R
\name{AddErrMsg}
\alias{AddErrMsg}
\title{Adds an error message}
\usage{
AddErrMsg(msg)

AddErrMsg(msg)
}
\arguments{
\item{msg}{Error message to print}
}
\description{
The error message will be printed in all cases.
Used in higher functions.

The error message will be printed in all cases.
Used in higher functions.
}
