% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboanalyst_enrichment.R
\name{CalculatePermHyperScore}
\alias{CalculatePermHyperScore}
\title{Over-representation analysis using hypergeometric tests}
\usage{
CalculatePermHyperScore(mSetObj = NA, model = T, perm = F, library.name)
}
\arguments{
\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Over-representation analysis using hypergeometric tests
The probability is calculated from obtaining equal or higher number
of hits using 1-phyper. Since phyper is a cumulative probability,
to get P(X>=hit.num) => P(X>(hit.num-1))
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
