% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_de_metabolites.R
\name{GetDefaultPvalCutoff}
\alias{GetDefaultPvalCutoff}
\title{Gets default p-val with reasonable cutoff}
\usage{
GetDefaultPvalCutoff(test.pvals)
}
\description{
Gets default p-val with reasonable cutoff
}
