% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_de_metabolites.R
\name{GetDEMetabolites}
\alias{GetDEMetabolites}
\title{Performs either tests of proportions or binomial tests
on the predicted metabolite table}
\usage{
GetDEMetabolites(mbSetObj, test.type = "proportions")
}
\arguments{
\item{test.type}{Either 'proportions' to do a test of proportions or
'binomial' to do a binomial test.}
}
\description{
Performs either tests of proportions or binomial tests
on the predicted metabolite table
}
