% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_de_metabolites.R
\name{GetMetabolitesFromGems}
\alias{GetMetabolitesFromGems}
\title{Function to grab metabolits from matched GEMs}
\usage{
GetMetabolitesFromGems(mbSetObj, nperm = 100)
}
\arguments{
\item{nperm}{Number of permutations, default is set to 100.}
}
\description{
This function creates a metabolite table
from matched GEMs as well as creates metabolite tables
for permutations (to be used by enrichment analysis).
Saves the permuted metabolite table in a list that is
saved as a .qs file
}
