% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{global_visualization}
\alias{global_visualization}
\title{PCA and t-SNE plots of predicted metabolomics data}
\usage{
global_visualization(mbSetObj, type = "pca", imgName = imgName)
}
\arguments{
\item{type}{Either 'pca' or 'tsne'.}

\item{imgName}{Character, the name of the plot without an extension.}
}
\description{
PCA and t-SNE plots of predicted metabolomics data
}
