% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{heatmap_bioactive_metabolites}
\alias{heatmap_bioactive_metabolites}
\title{Creates heatmap of bioactive compounds}
\usage{
heatmap_bioactive_metabolites(
  mbSetObj,
  metType = "bioactive",
  plotName = "heatmap",
  order.by = "hits",
  width = 7,
  height = 7
)
}
\arguments{
\item{metType}{Either 'bioactive' for bioactive metabolites or 'amino' for amino acids.}

\item{plotName}{Default is 'heatmap'.}

\item{order.by}{'Hits' to order by the number of hits or else it will order the metabolites alphabetically.}

\item{width}{Default is 7.}

\item{height}{Default is 7.}
}
\description{
Heatmap can be of either known microbial-origin
bioactive compounds or amino acids.
}
