% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{lollipop_chart_real_res}
\alias{lollipop_chart_real_res}
\title{Function to create lollipop comparison chart with asterisks
for significant pathways from real results.}
\usage{
lollipop_chart_real_res(
  mbSetObj,
  plot.libname,
  save.libname,
  asterisk_file,
  mum = F,
  width = 14,
  height = 7,
  lib.type = "kegg",
  anot = F,
  npath = 25
)
}
\arguments{
\item{asterisk_file}{Dataframe from csv containing enrichment results
from real omics data.}
}
\description{
Function to create lollipop comparison chart with asterisks
for significant pathways from real results.
}
