% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{lollipop_plot_comp}
\alias{lollipop_plot_comp}
\title{Function to create lollipop comparison chart
of pathway enrichment results from AGORA
versus CarveMe}
\usage{
lollipop_plot_comp(mbSetObj, plot.libname, save.libname, npath = 25)
}
\arguments{
\item{plot.libname}{'KEGG' or 'MetaCyc'.}

\item{save.libname}{Name of the pathway library to be used in the image name.}

\item{npath}{Default is set to 25.}
}
\description{
Function to create lollipop comparison chart
of pathway enrichment results from AGORA
versus CarveMe
}
