% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{lollipop_plot_enrichment}
\alias{lollipop_plot_enrichment}
\title{Function to create lollipop comparison chart
of pathway enrichment results from AGORA
versus CarveMe}
\usage{
lollipop_plot_enrichment(mbSetObj, plotName, anot = T, width = 7, height = 4)
}
\arguments{
\item{plotName}{Character, the name of the plot.}

\item{anot}{Add annotations for the pathways.}

\item{width}{Default is 7.}

\item{height}{Default is 4.}
}
\description{
Function to create lollipop comparison chart
of pathway enrichment results from AGORA
versus CarveMe
}
