% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_matching.R
\name{match2gems}
\alias{match2gems}
\title{This function is a wrapper function
to match user's input to our local
database of GEM models.}
\usage{
match2gems(mbSetObj, identity = 0.97)
}
\arguments{
\item{identity}{Numeric, only useful when matching ASVs. Default is 0.97 identity.}
}
\description{
Maps user's taxonomic
signatures to our database of GEM models.
}
