% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_plotting.R
\name{microbe_metabolite_bubble_chart}
\alias{microbe_metabolite_bubble_chart}
\title{Creates a bubble plot of bioactive
metabolite presence or absence
from a group of metabolites}
\usage{
microbe_metabolite_bubble_chart(
  mbSetObj,
  metType = "bioactive",
  width = 12,
  height = 7,
  bubble_size = 5
)
}
\arguments{
\item{metType}{Either 'bioactive' for bioactive metabolites, 'amino' for amino acids,
or the pathway name from the KEGG pathways.}

\item{width}{Default is 12.}

\item{height}{Default is 7.}

\item{bubble_size}{Default is 5.}
}
\description{
Creates a bubble plot of bioactive
metabolite presence or absence
from a group of metabolites
}
