% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_matching.R
\name{PerformMicrobiomeNetSanity}
\alias{PerformMicrobiomeNetSanity}
\title{This function does a sanity check on the user's
uploaded table}
\usage{
PerformMicrobiomeNetSanity(mbSetObj, userPheno)
}
\arguments{
\item{userPheno}{Character, input the user phenotype.}
}
\description{
This function does a sanity check on the user's
uploaded table
}
