% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_de_metabolites.R
\name{PerformPermEnrichment}
\alias{PerformPermEnrichment}
\title{Function that performs enrichment analysis
on permuted metabolites}
\usage{
PerformPermEnrichment(mbSetObj, library.name)
}
\arguments{
\item{library.name}{Input the name of the pathway library, without extension.}
}
\description{
Function that performs enrichment analysis
on permuted metabolites
}
