% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_de_metabolites.R
\name{PerformTrueEnrichment}
\alias{PerformTrueEnrichment}
\title{Wrapper function that performs enrichment analysis}
\usage{
PerformTrueEnrichment(mbSetObj, library.name, model = T, input.type = "name")
}
\arguments{
\item{library.name}{Input the name of the pathway library, without extension.}

\item{model}{Either 'both', 'agora', or 'carveme'.}
}
\description{
Wrapper function that performs enrichment analysis
}
