% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_graphics.R
\name{PlotCircularBubble}
\alias{PlotCircularBubble}
\title{Function to create circular bubble plots.}
\usage{
PlotCircularBubble(mbSet, "norm", "bubble", "Class", "OTU")
}
\arguments{
\item{mbSetObj}{Input the name of the mbSetObj.}

\item{dataInput}{Character, use "norm" to use the normalized relative abundance
or "filt" to use the filtered abundance.}

\item{bubbleplotName}{Character, input the preferred name of the circular
bubble plot to be created. Note that the name will be used with each group so if
you input "bubble", the plot will be saved as "bubble_group1.png".}

\item{facet}{Character, input the group label for the plot.}

\item{appendnm}{Boolean.}

\item{colpalopt}{Input whether the color palette will use R Color Brewer, "brewer",
or for users who wish to use a color palette robust to colorblindness
/(https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html/),
use "viridis".}

\item{colPalette}{Select the color palette options. "set3",
which is the Set3 from the R Color Brewer, "cont21" which is
a set of 21 colors, "cont28" which is a set of 28 colors, and
"cont42" which is a set of 42 colors.}

\item{fontSize}{Numeric, input the font size to be used to label the nodes
in the circular bubble plot. Default size is 2.5.}

\item{fontColor}{Character, input the color of the font to label the nodes. Default
color is set to "white".}

\item{format}{Character, input the preferred
format of the plot. By default it is set to "png".}

\item{dpi}{Numeric, input the dots per inch. By default
it is set to 72.}

\item{taxaLvl}{Input the name of the taxonomic level to view the circular bubble
plots.}
}
\description{
This functions creates a circular bubble plot of the
relative abundance of taxa in each group. Each group will have their own
plot outputted to the current working directory.
}
\author{
Jasmine Chong, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
