% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomenet_matching.R
\name{readMNetTable}
\alias{readMNetTable}
\title{This function reads in the user's data.}
\usage{
readMNetTable(mbSetObj, fileName, idType)
}
\arguments{
\item{fileName}{Character, input the name of the file containing the taxonomic signature.}

\item{idType}{Either 'asv', 'species', 'gg_otus', 'silva'.}
}
\description{
This function reads in the user's data.
}
